requirements.txt
# FastAPI and server
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic
pydantic-settings
python-multipart==0.0.6

# LangChain and AI
langchain==0.3.26
langchain-openai==0.2.0
openai==1.54.3
tiktoken==0.8.0

# Vector Database
pinecone==5.0.1

# Document Processing
pypdf==5.1.0
PyPDF2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.2

# Audio Processing
elevenlabs==0.2.27
pydub==0.25.1

# HTTP and utilities
requests==2.32.3
python-dotenv==1.0.1
